backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
google-cloud-aiplatform==1.49.0
PyMuPDF==1.24.1
gunicorn==22.0.0
google-cloud-vision